Create workflow backend component
