+++ wfbackend/wfbackend.go
@@ -1,34 +1,27 @@
 /*
-Copyright 2023 The Dapr Authors
+Copyright 2021 The Dapr Authors
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
-
-	http://www.apache.org/licenses/LICENSE-2.0
-
+    http://www.apache.org/licenses/LICENSE-2.0
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package workflows
 
-import (
-	"github.com/dapr/kit/logger"
-)
+package wfbackend
 
-func NewWorkflowBackendComp(logger logger.Logger) WorkflowBackend {
-	return &workflowBackendComponent{
-		logger: logger,
-	}
+import "github.com/dapr/components-contrib/metadata"
+
+// Workflow is an interface to perform operations on Workflow.
+type WorkflowBackend interface {
+	Init(metadata Metadata) error
 }
 
-type workflowBackendComponent struct {
-	logger logger.Logger
-}
-
-func (c *workflowBackendComponent) Init(metadata Metadata) error {
-	c.logger.Info("Initializing Dapr workflow backend component ")
-	return nil
+type WorkflowBackendComponentInfo struct {
+	WorkflowBackendType     string
+	WorkflowBackendMetadata metadata.Base
+	InvalidWorkflowBackend  bool
 }
